fix authorization

+++ app/Http/Controllers/StoryController.php
@@ -13,33 +13,39 @@ class StoryController extends Controller
 {
     public function __construct()
     {
-        if (Gate::denies('user-active')) {
-            abort(403);
-        }
         $this->middleware('auth');
+    }
+
+    public function index()
+    {
+        abort(404);
     }
 
     public function store(StoryRequest $request)
     {
-        $storyDataArray = array(
-            "categories_id" => $request->category,
-            "content" => $request->content,
-            "status" => $request->status,
-            "users_id" => Auth::id(),
-        );
-       
-        $story = Story::create($storyDataArray);
-        if ($request->photos != null) {
-            foreach ($request->photos as $photo) {
-                $newImageName = 'storage/image/' .uniqid() . '.' . $photo->extension();
-                $photo->move(public_path('storage/image'), $newImageName);
-                $story->images()->create([
-                   'image_url' => $newImageName,
-                ]);
+        if (Gate::allows('is-active')) {
+            $storyDataArray = array(
+                "categories_id" => $request->category,
+                "content" => $request->content,
+                "status" => $request->status,
+                "users_id" => Auth::id(),
+            );
+           
+            $story = Story::create($storyDataArray);
+            if ($request->photos != null) {
+                foreach ($request->photos as $photo) {
+                    $newImageName = 'storage/image/' .uniqid() . '.' . $photo->extension();
+                    $photo->move(public_path('storage/image'), $newImageName);
+                    $story->images()->create([
+                       'image_url' => $newImageName,
+                    ]);
+                }
             }
+    
+            return redirect()->route('home')->with('message', trans('message.create_success'));
+        } else {
+            abort(403);
         }
-
-        return redirect()->route('home')->with('message', trans('message.create_success'));
     }
 
     public function show($id)
@@ -53,9 +59,7 @@ class StoryController extends Controller
     public function edit($id)
     {
         $story = Story::findOrFail($id);
-        if (Gate::denies('story', $story)) {
-            abort(403);
-        }
+        $this->authorize('update', $story);
         $categories = Category::all();
 
         return view('homepage.story_edit', compact('story', 'categories'));
@@ -64,6 +68,8 @@ class StoryController extends Controller
     public function update(Request $request, $id)
     {
         $story = Story::findOrFail($id);
+        $this->authorize('update', $story);
+
         $story->update([
             'content' =>  $request->content,
             'status' => $request->status,
@@ -87,9 +93,8 @@ class StoryController extends Controller
     public function destroy($id)
     {
         $story = Story::findOrFail($id);
-        if (Gate::denies('story', $story)) {
-            abort(403);
-        }
+        $this->authorize('delete', $story);
+
         $story->images()->delete();
         Story::withTrashed()->where('id', $id)->forceDelete();
 
@@ -100,15 +105,19 @@ class StoryController extends Controller
 
     public function hideStory($id)
     {
-        $story = Story::findOrFail($id)->delete();
-        if ($story != null) {
-            return response()->json([
-                'message' => 'success',
-            ]);
+        if (Gate::allows('is-admin') || Gate::allows('is-inspector')) {
+            $story = Story::findOrFail($id)->delete();
+            if ($story != null) {
+                return response()->json([
+                    'message' => 'success',
+                ]);
+            } else {
+                return response()->json([
+                    'message' => 'fail',
+                ]);
+            }
         } else {
-            return response()->json([
-                'message' => 'fail',
-            ]);
+            abort(403);
         }
     }
 }
